move ragged ndarray function out of pandas and in general utils
