+++ python/ray/data/_internal/simple_block.py
@@ -5,6 +5,7 @@ from typing import Union, Callable, Iterator, List, Tuple, Any, Optional, TYPE_C
 
 import numpy as np
 
+from ray.air.util.tensor_extensions.utils import _create_possibly_ragged_ndarray
 
 if TYPE_CHECKING:
     import pandas
@@ -104,12 +105,6 @@ class SimpleBlockAccessor(BlockAccessor):
             if not isinstance(columns, list):
                 columns = [columns]
             return BlockAccessor.for_block(self.select(columns)).to_numpy()
-        # Import the utility function below only when needed
-        # in order to defer pyarrow import until runtime
-        from ray.air.util.tensor_extensions.pandas import (
-            _create_possibly_ragged_ndarray,
-        )
-
         return _create_possibly_ragged_ndarray(self._items)
 
     def to_arrow(self) -> "pyarrow.Table":
